Bring up WireGuard tunnel interface
Fix tests requiring the local WireGuard peer.
