+++ test/Cargo.toml
@@ -48,6 +48,7 @@ tokio = { version = "1.44", default-features = false }
 tokio-serial = "5.4.1"
 tonic = "0.13.1"
 tower = "0.5.1"
+tun = "0.8.8"
 windows-sys = "0.52.0"
 
 [workspace.lints.clippy]
