+++ test/test-manager/Cargo.toml
@@ -59,6 +59,7 @@ tokio-serial = { workspace = true }
 tokio-util = { version = "0.7", default-features = false, features = ["codec"] }
 tonic = { workspace = true }
 tower = { workspace = true }
+tun = { workspace = true }
 uuid = "1.3"
 
 [target.'cfg(target_os = "linux")'.dependencies]
