+++ test/test-manager/src/tests/helpers.rs
@@ -107,7 +107,7 @@ pub async fn reboot(rpc: &mut ServiceClient) -> Result<(), Error> {
     rpc.reboot().await?;
 
     #[cfg(target_os = "macos")]
-    crate::vm::network::macos::configure_tunnel()
+    crate::vm::network::wireguard::configure_tunnel()
         .await
         .context("Failed to recreate custom wg tun: {error}")?;
