+++ test/test-manager/src/vm/network/linux.rs
@@ -239,78 +239,6 @@ async fn start_dnsmasq() -> Result<DhcpProcHandle> {
     })
 }
 
-/*
-/// Creates a WireGuard peer on the host.
-///
-/// This relay does not support PQ handshakes, etc.
-///
-/// The client should connect to `CUSTOM_TUN_REMOTE_REAL_ADDR` on port `CUSTOM_TUN_REMOTE_REAL_PORT`
-/// using the private key `CUSTOM_TUN_LOCAL_PRIVKEY`, and tunnel IP `CUSTOM_TUN_LOCAL_TUN_ADDR`.
-///
-/// The public key of the peer is `CUSTOM_TUN_REMOTE_PUBKEY`. The tunnel IP of the host peer is
-/// `CUSTOM_TUN_REMOTE_TUN_ADDR`.
-async fn create_local_wireguard_peer() -> Result<()> {
-    run_ip_cmd([
-        "link",
-        "add",
-        "dev",
-        CUSTOM_TUN_INTERFACE_NAME,
-        "type",
-        "wireguard",
-    ])
-    .await?;
-    run_ip_cmd([
-        "addr",
-        "add",
-        "dev",
-        CUSTOM_TUN_INTERFACE_NAME,
-        &CUSTOM_TUN_REMOTE_TUN_ADDR.to_string(),
-        "peer",
-        &CUSTOM_TUN_LOCAL_TUN_ADDR.to_string(),
-    ])
-    .await?;
-
-    let mut tempfile = async_tempfile::TempFile::new()
-        .await
-        .map_err(Error::CreateWireguardConfig)?;
-
-    tempfile
-        .write_all(
-            format!(
-                "
-
-[Interface]
-PrivateKey = {CUSTOM_TUN_REMOTE_PRIVKEY}
-ListenPort = {CUSTOM_TUN_REMOTE_REAL_PORT}
-
-[Peer]
-PublicKey = {CUSTOM_TUN_LOCAL_PUBKEY}
-AllowedIPs = {CUSTOM_TUN_LOCAL_TUN_ADDR}
-
-"
-            )
-            .as_bytes(),
-        )
-        .await
-        .map_err(Error::WriteWireguardConfig)?;
-
-    let mut cmd = Command::new("wg");
-    cmd.args([
-        "setconf",
-        CUSTOM_TUN_INTERFACE_NAME,
-        tempfile.file_path().to_str().unwrap(),
-    ]);
-    let output = cmd.output().await.map_err(Error::WgStart)?;
-    if !output.status.success() {
-        return Err(Error::WgFailed(output.status.code().unwrap()));
-    }
-
-    run_ip_cmd(["link", "set", "dev", CUSTOM_TUN_INTERFACE_NAME, "up"]).await?;
-
-    Ok(())
-}
-*/
-
 async fn run_ip_cmd<I, S>(args: I) -> Result<()>
 where
     I: IntoIterator<Item = S>,
