+++ test/test-manager/src/vm/network/wireguard.rs
@@ -4,7 +4,6 @@ use anyhow::{Context, Result};
 use gotatun::device::{self, DefaultDeviceTransports, Device, Peer};
 use gotatun::tun::tun_async_device::TunDevice;
 use ipnetwork::Ipv4Network;
-
 use std::net::Ipv4Addr;
 
 // Private key of the wireguard remote peer on host.
@@ -33,7 +32,10 @@ pub const CUSTOM_TUN_REMOTE_TUN_ADDR: Ipv4Addr = Ipv4Addr::new(192, 168, 15, 1);
 /// Gateway (and default DNS resolver) of the wireguard tunnel.
 pub const CUSTOM_TUN_GATEWAY: Ipv4Addr = CUSTOM_TUN_REMOTE_TUN_ADDR;
 /// Name of the wireguard interface on the host
-pub const CUSTOM_TUN_INTERFACE_NAME: &str = "wg-relay0";
+pub const CUSTOM_TUN_INTERFACE_NAME: &str = cfg_select! {
+    target_os = "linux" => { "wg-relay0" }
+    target_os = "macos" => { "utun123" }
+};
 
 /// Creates a WireGuard peer on the host.
 ///
@@ -51,8 +53,19 @@ pub(crate) async fn create_interface() -> Result<Device<DefaultDeviceTransports>
         const { Ipv4Network::new_checked(CUSTOM_TUN_LOCAL_TUN_ADDR, 32).unwrap() }.into(),
     );
 
-    let tun =
-        TunDevice::from_name(CUSTOM_TUN_INTERFACE_NAME).context("Failed to create utun device")?;
+    let tun = {
+        let tun = {
+            let mut tun_config = tun::Configuration::default();
+            tun_config.tun_name(CUSTOM_TUN_INTERFACE_NAME);
+            tun_config.up();
+            #[cfg(target_os = "macos")]
+            tun_config.platform_config(|p| {
+                p.enable_routing(false);
+            });
+            tun::create_as_async(&tun_config).context("Failed to open tun device")?
+        };
+        TunDevice::from_tun_device(tun).context("Failed to create utun device")?
+    };
 
     let device = device::build()
         .with_default_udp()
@@ -65,4 +78,36 @@ pub(crate) async fn create_interface() -> Result<Device<DefaultDeviceTransports>
         .context("Failed to create gotatun device")?;
 
     Ok(device)
+}
+
+#[cfg(target_os = "macos")]
+pub async fn configure_tunnel() -> Result<()> {
+    use anyhow::bail;
+    use tokio::process::Command;
+    // Check if the tunnel device is configured
+    let mut cmd = Command::new("/usr/sbin/ipconfig");
+    cmd.args(["getifaddr", CUSTOM_TUN_INTERFACE_NAME]);
+    let output = cmd
+        .output()
+        .await
+        .context("Check if wireguard tunnel has IP")?;
+    if output.status.success() {
+        log::debug!("Tunnel {CUSTOM_TUN_INTERFACE_NAME} already configured");
+        return Ok(());
+    }
+
+    // Set tunnel IP address
+    let mut cmd = Command::new("/usr/bin/sudo");
+    cmd.args([
+        "/usr/sbin/ipconfig",
+        "set",
+        CUSTOM_TUN_INTERFACE_NAME,
+        "manual",
+        &CUSTOM_TUN_REMOTE_TUN_ADDR.to_string(),
+    ]);
+    let status = cmd.status().await.context("Run ipconfig")?;
+    if !status.success() {
+        bail!("ipconfig failed: {}", status.code().unwrap());
+    }
+    Ok(())
 }
